chore: rename plugin-router-legacy to plugin-router-v5
