+++ packages/runtime/plugin-garfish/src/runtime/utils/apps.tsx
@@ -3,7 +3,7 @@ import path from 'path';
 import React from 'react';
 // eslint-disable-next-line import/no-named-as-default
 import Garfish, { interfaces } from 'garfish';
-import { withRouter } from '@modern-js/plugin-router-legacy/runtime';
+import { withRouter } from '@modern-js/plugin-router-v5/runtime';
 // import Loadable from 'react-loadable';
 import { Manifest, MicroComponentProps, ModulesInfo } from '../useModuleApps';
 import { logger, generateSubAppContainerKey } from '../../util';
