+++ packages/runtime/plugin-router-v5/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@modern-js/plugin-router-legacy",
+  "name": "@modern-js/plugin-router-v5",
   "description": "The meta-framework suite designed from scratch for frontend-focused modern web development.",
   "homepage": "https://modernjs.dev",
   "bugs": "https://github.com/modern-js-dev/modern.js/issues",
