+++ packages/runtime/plugin-router-v5/src/modern-app-env.d.ts
