+++ packages/solutions/app-tools/src/index.ts
@@ -112,7 +112,7 @@ export default (): CliPlugin<AppTools> => ({
     '@modern-js/plugin-document',
     '@modern-js/plugin-state',
     '@modern-js/plugin-router',
-    '@modern-js/plugin-router-legacy',
+    '@modern-js/plugin-router-v5',
     '@modern-js/plugin-polyfill',
   ],
