+++ tests/integration/app-document/src/modern-app-env.d.ts
@@ -1,3 +1,3 @@
 /// <reference types='@modern-js/app-tools/types' />
 /// <reference types='@modern-js/runtime/types' />
-/// <reference types='@modern-js/plugin-router-legacy/types' />
+/// <reference types='@modern-js/plugin-router-v5/types' />
