+++ tests/integration/garfish/dashboard/package.json
@@ -24,7 +24,7 @@
   },
   "devDependencies": {
     "@modern-js/app-tools": "workspace:*",
-    "@modern-js/plugin-router-legacy": "workspace:*",
+    "@modern-js/plugin-router-v5": "workspace:*",
     "@types/node": "^14",
     "@types/react": "^18",
     "@types/react-dom": "^18",
