+++ tests/integration/garfish/main/package.json
@@ -24,7 +24,7 @@
   },
   "devDependencies": {
     "@modern-js/app-tools": "workspace:*",
-    "@modern-js/plugin-router-legacy": "workspace:*",
+    "@modern-js/plugin-router-v5": "workspace:*",
     "typescript": "^4",
     "@types/react": "^18",
     "@types/react-dom": "^18",
