+++ tests/integration/garfish/main/src/modern-app-env.d.ts
@@ -3,4 +3,4 @@
 /// <reference types='@modern-js/plugin-garfish/type' />
 /// <reference types='@modern-js/plugin-express/types' />
 /// <reference types='@modern-js/plugin-koa/types' />
-/// <reference types='@modern-js/plugin-router-legacy/types' />
+/// <reference types='@modern-js/plugin-router-v5/types' />
