+++ tests/integration/runtime/fixtures/file-based-router/modern.config.ts
@@ -1,5 +1,5 @@
 import AppToolsPlugin, { defineConfig } from '@modern-js/app-tools';
-import RouterLegacyPlugin from '@modern-js/plugin-router-legacy';
+import RouterLegacyPlugin from '@modern-js/plugin-router-v5';
 
 export default defineConfig({
   runtime: {
